+++ graphql/mutations/createContactQuery.js
@@ -2,6 +2,7 @@ import { gql } from "@apollo/client";
 
 export const CREATE_CONTACT_QUERY = gql`
   mutation CreateContact(
+    $teamId: String
     $name: String!
     $email: String!
     $phone: String!
@@ -9,12 +10,14 @@ export const CREATE_CONTACT_QUERY = gql`
     $message: String!
   ) {
     createContact(
+      teamID: $teamId
       name: $name
       email: $email
       phone: $phone
       subject: $subject
       message: $message
     ) {
+      teamID
       name
       email
       phone