+++ graphql/queries/GetGeneralContacts.js
@@ -3,6 +3,7 @@ import { gql } from "@apollo/client";
 export const GET_GENERAL_QUERIES = gql`
   query GetContacts($value: String!, $limNum: Int) {
     getContacts(value: $value, lim_num: $limNum) {
+      teamID
       name
       email
       phone